+++ src/assets/AnchorLinkIcon/index.ts
@@ -1,0 +1,3 @@
+import { AnchorLinkIcon } from "./AnchorLinkIcon";
+
+export default AnchorLinkIcon;