+++ src/assets/ChevronLeftIcon/ChevronLeftIcon.tsx
@@ -1,7 +1,7 @@
 import { IconProps } from "../../types/shared";
 
 //Icon Link: https://www.svgrepo.com/svg/533659/chevron-left
-const ChevronLeftIcon = ({ width, height, color }: IconProps) => {
+const ChevronLeftIcon = ({ width, height, color, className }: IconProps) => {
   return (
     <svg
       width={width ? width : 32}
@@ -9,6 +9,7 @@ const ChevronLeftIcon = ({ width, height, color }: IconProps) => {
       viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      className={`chevron-left-icon ${className}`}
     >
       <path
         d="M15 6L9 12L15 18"