+++ src/assets/ChevronUpIcon/ChevronUpIcon.tsx
@@ -1,7 +1,7 @@
 import { IconProps } from "../../types/shared";
 
 //Icon Link: https://www.svgrepo.com/svg/533664/chevron-up
-const ChevronUpIcon = ({ width, height, color }: IconProps) => {
+const ChevronUpIcon = ({ width, height, color, className }: IconProps) => {
   return (
     <svg
       width={width ? width : 32}
@@ -9,6 +9,7 @@ const ChevronUpIcon = ({ width, height, color }: IconProps) => {
       viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      className={`chevron-up-icon ${className}`}
     >
       <path
         d="M6 15L12 9L18 15"