+++ src/assets/HalfHeartFilledIcon/index.ts
@@ -1,0 +1,3 @@
+import { HalfHeartFilledIcon } from "./HalfHeartFilledIcon";
+
+export default HalfHeartFilledIcon;