+++ src/assets/HomeIcon/HomeIcon.tsx
@@ -1,0 +1,22 @@
+import { IconProps } from "../../types/shared";
+
+//Icon Link: https://www.svgrepo.com/svg/535437/home
+const HomeIcon = ({ width, height, color, className }: IconProps) => {
+  return (
+    <svg
+      width={width ? width : 32}
+      height={height ? height : 32}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={`home-icon ${className}`}
+    >
+      <path
+        d="M1 6V15H6V11C6 9.89543 6.89543 9 8 9C9.10457 9 10 9.89543 10 11V15H15V6L8 0L1 6Z"
+        fill={color ? color : "#000000"}
+      />
+    </svg>
+  );
+};
+
+export { HomeIcon };