+++ src/assets/HomeVariantOneIcon/index.ts
@@ -1,0 +1,3 @@
+import { HomeVariantOneIcon } from "./HomeVariantOneIcon";
+
+export default HomeVariantOneIcon;