+++ src/assets/InstagramOldIcon/index.ts
@@ -1,0 +1,3 @@
+import { InstagramOldIcon } from "./InstagramOldIcon";
+
+export default InstagramOldIcon;