+++ src/assets/LocationVariantOneIcon/LocationVariantOneIcon.tsx
@@ -1,0 +1,30 @@
+import { IconProps } from "../../types/shared";
+
+//Icon Link: https://www.svgrepo.com/svg/532544/location-pin-alt
+const LocationVariantOneIcon = ({
+  width,
+  height,
+  color,
+  className,
+}: IconProps) => {
+  return (
+    <svg
+      width={width ? width : 32}
+      height={height ? height : 32}
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={`location-variation-1-icon ${className}`}
+    >
+      <path
+        d="M12 21C15.5 17.4 19 14.1764 19 10.2C19 6.22355 15.866 3 12 3C8.13401 3 5 6.22355 5 10.2C5 14.1764 8.5 17.4 12 21Z"
+        stroke={color ? color : "#000000"}
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+
+export { LocationVariantOneIcon };