+++ src/assets/LocationVariantOneIcon/index.ts
@@ -1,0 +1,3 @@
+import { LocationVariantOneIcon } from "./LocationVariantOneIcon";
+
+export default LocationVariantOneIcon;