+++ src/assets/MessengerIcon/index.ts
@@ -1,0 +1,3 @@
+import { MessengerIcon } from "./MessengerIcon";
+
+export default MessengerIcon;