+++ src/assets/ProfileVariantOneIcon/index.ts
@@ -1,0 +1,3 @@
+import { ProfileVariantOneIcon } from "./ProfileVariantOneIcon";
+
+export default ProfileVariantOneIcon;