+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [0.1.0] - 2025-02-10
+
+### Added
+
+- Implemented Anchor Link, Bag, Bag Variant 1, Bag Variant 2, Cart, Cart Variant 1, Cart Variant 2, Facebook, Half Heart Filled, Half Heart, Half Star Filled, Half Star, Heart Filled, Heart, Home, Home Variant 1, Instagram, Instagram Old, Location, Location Variant 1, Menu, Messenger, Profile, Profile Variant 1, Search, Star Filled, Star, Twitter and Whatsapp Icons, implemented className prop, rewrite of README file.
+
 ## [0.0.2] - 2025-02-09
 
 ### Added
 
-- Implementing Chevron Right and Left icons and removing "any" types.
+- Implemented Chevron Right and Left icons and removing "any" types.
 
 ## [0.0.1] - 2025-02-09